revised navigators
