+++ screens/termsOfService.js
@@ -20,7 +20,6 @@ export default function termsOfService() {
     <SafeAreaView>
       <ScrollView style={styles.scrollView}>
         <Text style={styles.text}>{`
-利用規約\n
 　この利用規約（以下，「本規約」といいます。）は，開発者（以下，「私」といいます。）がこのアプリケーション上で提供するサービス（以下，「本サービス」といいます。）の利用条件を定めるものです。利用者の皆さま（以下，「ユーザー」といいます。）には，本規約に従って，本サービスをご利用いただきます。\n
 
 第1条（適用）\n
